+++ asgard-run.sh
@@ -15,6 +15,7 @@ sleep 3
 
 source "dev/network.sh"
 source "dev/pgsql.sh"
+source "dev/elasticsearch.sh"
 source "dev/zk.sh"
 source "dev/marathon.sh"
 source "dev/mesosmaster.sh"