+++ asgard/backends/mesos/models/app.py
@@ -3,7 +3,6 @@ from asgard.models.app import App
 
 class MesosApp(App):
     type: str = "MESOS"
-    id: str
 
     def transform_to_asgard_app_id(executor_id: str) -> str:
         task_name_part = executor_id.split(".")[0]