+++ asgard/services/apps.py
@@ -1,0 +1,14 @@
+from asgard.backends.base import Orchestrator
+from asgard.backends.mesos.models.app import MesosApp
+from asgard.models.account import Account
+from asgard.models.app import AppStats
+from asgard.models.user import User
+
+
+class AppsService:
+    @staticmethod
+    async def get_app_stats(
+        app_id: str, user: User, account: Account, orchestrator: Orchestrator
+    ) -> AppStats:
+        app = MesosApp(id=app_id)
+        return await orchestrator.get_app_stats(app, user, account)