+++ itests/util.py
@@ -1,3 +1,5 @@
+import asyncio
+import os
 import random
 import string
 from collections import defaultdict
@@ -5,6 +7,7 @@ from importlib import reload
 from typing import Any, Dict, List, Type, Set
 
 import asyncworker
+from aioelasticsearch import Elasticsearch
 from aiohttp import web
 from aiohttp.test_utils import TestClient, TestServer
 from aiopg.sa import Engine
@@ -168,6 +171,7 @@ class BaseTestCase(TestCase):
 
     async def setUp(self):
 
+        self.esclient = Elasticsearch([settings.STATS_API_URL])
         self.pg_data_mocker = PgDataMocker(pool=await self.conn_pool())
         self.users_fixture = [
             [
@@ -233,14 +237,33 @@ class BaseTestCase(TestCase):
 
     async def tearDown(self):
         await self.pg_data_mocker.drop()
+        if hasattr(self, "server"):
+            await self.server.close()
 
     async def aiohttp_client(self, app: asyncworker.App) -> TestClient:
         routes = app.routes_registry.http_routes
         http_app = web.Application()
         for route in routes:
             http_app.router.add_route(**route)
-        server = TestServer(http_app)
-        client = TestClient(server)
-        await server.start_server()
+        self.server = TestServer(
+            http_app, port=os.environ["TEST_ASYNCWORKER_HTTP_PORT"]
+        )
+        client = TestClient(self.server)
+        await self.server.start_server()
 
         return client
+
+    async def _load_app_stats_into_storage(
+        self, index_name, timestamp_to_use, datapoints
+    ):
+        """
+        Carrega no elasticsearch local os datapoints de estatísticas de apps
+        O `timestamp` usado em cada datapoint é o `timestamp_to_use`.
+        """
+        for datapoint in datapoints:
+            datapoint["timestamp"] = timestamp_to_use.isoformat()
+            await self.esclient.index(
+                index=index_name, doc_type="stats", body=datapoint
+            )
+        # para dar tempo do Elasticsearch local indexar os dados recém inseridos
+        await asyncio.sleep(1)