+++ tests/__init__.py
@@ -18,14 +18,3 @@ class RequestStub(object):
 
     def get_json(self):
         return json.loads(self.data)
-
-
-class ContextStub(object):
-    def __init__(self, marathon_client=None):
-        self.marathon_client = marathon_client
-
-
-def rebuild_schema():
-    assert engine.name == "sqlite"
-    BaseModelAlchemy.metadata.drop_all(engine)
-    BaseModelAlchemy.metadata.create_all(engine)